Create SKOSGraphDataModelTaxonomy.tsl
The TSL (Trinity Specification Language) code defines four cells and two structs representing different components of the Simple Knowledge Organization System (SKOS).

SKOSConcept cell: Represents an individual concept in the SKOS system. It contains the following fields:

Name, Notation, ConceptScheme, TopConceptOf, and various lists of LanguageLabel and LanguageNote for different purposes (PrefLabel, AltLabel, HiddenLabel, Definition, ScopeNote, Example, HistoryNote, EditorialNote, and ChangeNote).
HyperEdges: A list of CellId, referring to related hyperedges.
SKOSConceptScheme cell: Represents a concept scheme in the SKOS system. It contains the following fields:

Name, Title, Creator, Contributor, Publisher, Rights, License, and Language.
Created and Modified: DateTime fields representing the creation and modification timestamps.
Concepts and TopConcepts: Lists of CellId, referring to related concepts and top concepts in the scheme.
SKOSHyperEdge cell: Represents a hyperedge in the SKOS system. It contains the following fields:

Relationship: A string representing the relationship between concepts.
RelatedConcepts: A list of CellId, referring to related concepts in the hyperedge.
SKOSHyperGraph cell: Represents a hypergraph in the SKOS system. It contains the following fields:

Name: A string representing the name of the hypergraph.
ConceptSchemes, Concepts, and HyperEdges: Lists of CellId, referring to related concept schemes, concepts, and hyperedges in the hypergraph.
LanguageLabel struct: Represents a language-specific label. It contains the following fields:

Language: A string representing the language code.
Label: A string representing the label text in the specified language.
LanguageNote struct: Represents a language-specific note. It contains the following fields:

Language: A string representing the language code.
Note: A string representing the note text in the specified language.
The [GraphNode] attribute applied to the cells indicates that they are part of a graph data model, and the [GraphEdge] attribute applied to some fields in the cells signifies that those fields represent edges in the graph connecting different nodes (cells).