How to model PropertyTypes and HyperGraph types
In this TSL code, we define two cell types: PropertyGraphNode and PropertyGraphEdge. The PropertyGraphNode cell type has two fields: Id and Properties. The Id field is of type CellId and represents the unique identifier of the node, while the Properties field is of type map<string, string> and stores the key-value pairs of properties associated with the node. Similarly, the PropertyGraphEdge cell type has two fields: Source, Destination, and Properties. The Source and Destination fields are of type CellId and represent the nodes connected by the edge, while the Properties field is a map that stores the properties of the edge.

In this TSL code, we define two cell types: HyperGraphCell and HyperGraphEdge. The HyperGraphCell cell type has two fields: Id and Nodes. The Id field is of type CellId and represents the unique identifier of the hypercell, while the Nodes field is a list of CellId values that represent the nodes connected by the hypercell. Similarly, the HyperGraphEdge cell type has two fields: Id and Nodes. The Id field is of type CellId and represents the unique identifier of the hyperedge, while the Nodes field is a list of CellId values that represent the nodes connected by the hyperedge.

Note that this TSL code defines a simple hypergraph, where each hypercell connects a set of nodes. More complex hypergraphs can be defined using different types of hyperedges or by defining higher-order relationships between hypercells.